Goal Plugins
If the `action_type` field in the map is set to `package/Goal` then toponav will set the
args of the goal at run time using the python function `Goal` imported from `package/src/Goal.py`.
Goal args (e.g. `continue_action` in ThorvaldNavGoal) can be set dynamically.
Setting the goal in the map is no longer a requirement.
If the plugin doesnt exist then the goal from the map is used.